+++ src/data/CommandTree1.16.ts
@@ -1220,7 +1220,7 @@ export const CommandTree: ICommandTree = {
                                             parser: new NumberArgumentParser('float', 0),
                                             children: {
                                                 count: {
-                                                    parser: new NumberArgumentParser('float', 0),
+                                                    parser: new NumberArgumentParser('integer', 0),
                                                     executable: true,
                                                     children: {
                                                         mode: {
